+++ src/main/resources/query_by_wait.sql
@@ -1,0 +1,3 @@
+SELECT wait_event, wait_event_type, count(*) as count
+FROM pg_stat_activity
+group by wait_event, wait_event_type